Outlined plan for independent motor control.
Essentially this meant treating each motor/wheel pair as an independent entity as opposed to balancing between the ex and y axis then balancing within each axis.

The major benefit here is there is no need for balancing if the wheel tracking is accurate then as the distance increases for one side but not the other that side will increase it's speed to compensate.