Fixed motor initialisation. See full message for gotchas...
First, there was no call to the begin() method of the motor driver's library.

The motor stucture used was a local copy due to the type being specified (again)
so the global "motors" struct was still empty.

A subtle problem was with the AFMS is that it cannot be defined locally.
Even though we only use it within the method, c++ is smart enough to know about
scope and lifetime. As a local object, there is an implicit call the the
destructor upon returning from the function. This totally undoes all of the
initialisation. Fix was to make it global.