Rewrote control module in a more data-driven way
Kept the algorithm essentially the same, but structed it around a struct
of motor state. Also treating speeds as a vector.

This means we get rid of the 4-way (And even 8-way!) conditionals and
can just loop through a collection of motor structs...

Still not perfect, but will be easier to debug and maintain